requirements.txt
boto3==1.26.0
moto==4.0.0
python-dotenv==1.0.0